Compress image, use cdn, optimize code
