+++ index.html
@@ -14,9 +14,10 @@
     <link rel="stylesheet" href="https://cdn.staticfile.org/OwlCarousel2/2.3.1/assets/owl.carousel.min.css">
     <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Open+Sans:300,400,600,700,800">
 
+    <link rel="stylesheet" href="./css/photo.css">
+    <link rel="stylesheet" href="./css/templatemo-style.css">
     <script src="js/vendor/modernizr-2.8.3-respond-1.4.2.min.js"></script>
-    <script src="https://cdn.staticfile.org/jquery/3.3.1/jquery.min.js"></script>
-    <script src="./dist/bundle.js"></script>
+    
     <title>CLFStudio</title>
 </head>
 
@@ -524,8 +525,11 @@
         </section>
     </div>
 
+    <script src="https://cdn.staticfile.org/jquery/3.3.1/jquery.min.js"></script>
     <script src="https://cdn.staticfile.org/twitter-bootstrap/3.3.1/js/bootstrap.min.js"></script>
     <script src="js/plugins.js"></script>
+    <script src="js/main.js"></script>
+    <script src="index.js"></script>
 </body>
 
 </html>
