+++ package.json
@@ -6,17 +6,10 @@
     "author": "tjx666 <2713151713@qq.com>",
     "license": "MIT",
     "scripts": {
-        "build": "eslint ./js --fix && webpack --config webpack.config.js --mode production"
+        "lint": "eslint . --fix"
     },
     "devDependencies": {
-        "css-loader": "^1.0.0",
-        "file-loader": "^2.0.0",
-        "fs-extra": "^7.0.0",
-        "style-loader": "^0.23.1",
-        "tinify": "^1.6.0-beta.2",
-        "url-loader": "^1.1.2",
-        "webpack": "^4.20.2",
-        "webpack-cli": "^3.1.2"
+       
     },
     "dependencies": {}
 }
