--- webpack.config.js
@@ -1,30 +1,0 @@
-const path = require('path');
-
-module.exports = {
-    entry: './index.js',
-    output: {
-        filename: 'bundle.js',
-        path: path.resolve(__dirname, 'dist')
-    },
-    module: {
-        rules: [{
-                test: /\.css$/,
-                use: [
-                    'style-loader',
-                    'css-loader'
-                ]
-            },
-            {
-                test: /\.(png|svg|jpg|gif)$/,
-                use: {
-                    loader: 'url-loader',
-                    options: {
-                        name: "imgs/[name].[hash].[ext]",
-                    }
-                }
-            }
-        ]
-    },
-    plugins: [
-      ]
-};
